add config cli api
